Implement proper cursor positioning
The cursor's movement is now restricted to within the bounds of the file
it current has in the buffer. A cursor can move along a lines content
until it reaches the start or end of the line. Similarly it can move up
an down through the buffers lines until it reaches the start or end of
the buffer.