Make editor wait for events before redraw
Changed the events channel to use recv instead of try_recv which means
that the draw loop will wait for an event before continuing to the next
draw iteration. This greatly reduces the number of draws happening per
second. If any event locking is happening in the future, though, this
change *could* be what introduced it.

This is related to issue #4 in terms of the performance aspect.