Fixing Iota to be completely thread safe.
This resolves double-free assertions that were happening consitently for
me in RustBox when I exited.  It also preserves stacktraces properly
without terminating RustBox early.

This is motivated by changes in RustBox (to replace the old unsafe API
with a safe Rust one).  Note that in theory, we should be able to split
RustBox out into two components, an input component and an output
component, but to do that properly and keep it threadsafe requires &mut
references to be sent across thread boundaries.  Currently that would
require unsafe code, so for the moment we revert to a single-threaded
solution.