Implemented Overlay
An overlay is an interface for interacting with the user. It captures
keyboard input and sends commands to the current mode which interprets
the input and acts on it.

In the future, we may need to move the overlay from the Editor to the
View. This would allow us to more easily set overlays from the modes,
rather than having this back-and-forth between the overlay, editor and
mode.

refs #64