Move overlay from editor to view
Having the overlay on the view is much cleaner as we can remove the
different branches in editor which were handling the response from the
overlay. Now, editor just handles the response from mode, which in turn
can pass responses from the view, or the overlay. Also, because mode
has access to the current view, its now much easier to set overlays on
the view from the current mode.