Refactor Buffer::Direction
Refactored the Direction enum to no longer contain uints. This uint was
used to determine the amount of operations to execute in the given
direction. I've changed this amount to be an explicit parameter to all
the places this was previously used. For instance, instead of calling
view.move_cursor(Direction::Left(1)) you now call
view.move_cursor(Direction::Left, 1) which is think is more idiomatic.

This will also make it easier to integration with the concept of a
TextObject, where we will be able to write something like:
view.move_cursor(Direction::Left, TextObject::Char) to move left by a
single char, for example.