Start experimenting with events system
This is an idea I'm playing with for overhauling the key event / command
resolution system. In the current system, it would be extremely
difficult to allow plugins or extensions to bind callbacks to keys, and
bind callbacks to editor events.

I want to allow plugins to intercept almost every event inside the
editor. This will allow any default behaviour to be customised by
extensions.

One of my plans here is to remove the modal editing that is present now,
and instead implement that as a plugin. To do this, plugins will need to
be able to:

- register and fire custom events
- intercept & respond to core events
- override core events

So, this is a start on implementing some sort of event system in iota.
It might work out, it might not. Let's see...