Refactor server structure
This adds the concept (or the start of the concept) of workspaces.
Workspace are just named groupings of buffers.

I've also fixed up the internal response format for the api.

And started making the python test script a bit better. Saves me
re-writing it every time I want to do something