Load themes from sublime text Base16 if present
If the Base16 package is not installed, the default syntect themes will
be used. It might be worth in the future trying to load all sublime
themes, but for now I'm going with Base16 as those are the ones I use
myself.

A more feasible long-term solution should be added later.