Make rustbox event polling block
This significantly reduces the CPU usage for iota, as it no longer has
to write to stdout multiple times a second.

In the future it may be worth looking into how to make it so that we
don't have to call "clear" during a draw. It might be that we can just
clear past the end of a line and clear lines that don't have content.

For now, this will do.