Removed the UIBuffer object
This didn't add anything over just writing directly to RustBox, and
there's no performance hit from doing so, so I'm removing the use of the
UIBuffer entirely.

If it turns out later that its a good idea it can be added back in,
however for now all it does is add unnecessary complexity.