Temporarily remove syntax highlighting
I think it was a little premature to add syntax highlighting here. There
are still many issues with the basic editing functionality that should
be addressed before tackling syntax highlighting.

I have lots of things on my list to improve in iota such as messaging,
the concept of overlays, modal editing, keybindings, and more. I
consider many of these to be of higher importance than syntax
highlighting right now, so I've decided to strip that out to simplify
the codebase to make tackling the other items that bit easier.

Syntax highlighting is definitly something I want for iota, and it will
certainly be re-added. When it is re-added I would like it to be far
more robust and extensible than it was in this iteration.