Begin work on overhauling the overlay system
This commit changes Overlay to a trait and removes all but one
implementation - the command prompt. Lots more work planned for this,
but this is the first step.

This also changes the overlay key handling to use the event builder
system which is more inline with how the rest of the editor handles
events.