Get the build passing again
Many keybindings are still broken however. The next patch needs to
change the keybinding type. At the moment we just store the name of the
command to be run when the key is pressed. We need to change this to
store the name of the command and also any arguments.