More work on the KeyBinding system
Still lots to do. Specifically:

It would be better to remove the keys field from the KeyBinding
struct and replace with something like:

    struct KeyBinding {
        keys: Vec<Key>,
        command_info: CommandInfo
    }

    struct CommandInfo {
        command_name: String,
        args: BuilderArgs
    }

This would be more efficient in the keymap (I think) and also remove
the need to us to define the keys in the example below when it
doesn't really make sense to do so.

Additionally, need to get all commands re-implemented. Again.