Refactor keymap to allow arguments in bindings
This change allows arguments to be given to key bindings and commands in
general.

Keys are now bound to CommandInfo objects which consist of a command
name and a set of arguments for that command. When a key is pressed we
check the bound command name against the set of all editor commands to
get the corresponding command-producing function. This function when
called with the associated arguments from the keybinding will produce a
Command for the editor to execute.

Ultimately all this work is leading toward being able to define
keybindings in JSON and have them deserialized into a KeyBinding and
CommandInfo pair.

For instance, the idea I was working off looked something like:
takes several argumnets:

    keys: ["down"]
    command: "buffer::move_cursor"
    args: {
        kind: "line"
        offset: "forward"
        number: 1
    }

The above would move the cursor down one line. The same pattern could be
used for all movement variations. This is obviously not valid JSON, but
it gives the idea of what I'm working towards.

Additionally, this change removes the requirement for Commands to be
Copy which was limiting the types of data that could be passed around in
a Command.